requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
bcrypt>=4.0.0
google-genai>=0.1.0
